Restructures item/request component/table to consistent format
