+++ src/Request/RequestComponent.js
@@ -1,0 +1,17 @@
+import React from "react";
+import RequestTable from "./RequestTable";
+
+
+class RequestComponent extends React.Component {
+
+    render() {
+      return (
+        <div>
+        <RequestTable />
+        </div>
+      )
+    }
+
+}
+
+export default RequestComponent;
